back-end-Admin_Sidebar

+++ Code/mern-website/admin-app/src/actions/auth.actions.js
@@ -60,9 +60,17 @@ export const isUserLoggedIn = () => {
 
 export const signout = () => {
     return async dispatch => {
-        localStorage.clear();
-        dispatch({
-            type: authConstants.LOGOUT_REQUEST
-        });
+        dispatch({ type: authConstants.LOGOUT_REQUEST});
+        const res = await axios.post("/admin/signout");
+        if (res.status === 200) {
+            localStorage.clear();
+        dispatch({ type: authConstants.LOGOUT_SUCCESS});
+        }
+        else{
+            dispatch({
+                type: authConstants.LOGOUT_FAILURE,
+                payload:{ error: res.data.error }
+            }); 
+        }        
     }
 }
